Aula 11 - Utilizando o useHistory para tratamento de erro

+++ src/pages/Repositories/index.js
@@ -1,24 +1,30 @@
 import React, { useState, useEffect } from 'react';
+import { useHistory } from 'react-router';
 
 import * as S from './styled';
 
 export default function Repositories() {
+    const history = useHistory();
+
     const [ repositories, setRepositories ] = useState([]);
 
     // useEffect monitora o estado de uma variável. Quando houver alteração, irá disparar uma função.
     useEffect( () => {
         let aData = localStorage.getItem('data'); //pega as informações que estão na Storage e armazena na variável.
 
-        //Converte o array em objeto.
-        let oData = JSON.parse(aData);
-        console.log(oData);
-        console.log(repositories);
+        if(aData != null) {
+            //Converte o array em objeto.
+            let oData = JSON.parse(aData);
+            console.log(oData);
+            console.log(repositories);
 
-        //Armazena o objeto no repositório de nomes
-        setRepositories(oData);
+            //Armazena o objeto no repositório de nomes
+            setRepositories(oData);
 
-        localStorage.clear();
-
+            localStorage.clear();
+        } else {
+            history.push("/"); //retorna para a pagina inicial
+        }
     }, []); //Se deixar os [] vazio, ao renderizar a página, a função será executada.
 
     return (
